Refactored Microsoft.Extension.* depedent code into Finbuckle.MultiTenant.Core.
- Updated unit tests for Core.Tests and EntityFrameworkCore.Tests
- AspNetCore projects still pending
