--- src/Finbuckle.MultiTenant.AspNetCore/InMemoryStoreOptions.cs
@@ -1,37 +1,0 @@
-//    Copyright 2018 Andrew White
-// 
-//    Licensed under the Apache License, Version 2.0 (the "License");
-//    you may not use this file except in compliance with the License.
-//    You may obtain a copy of the License at
-// 
-//        http://www.apache.org/licenses/LICENSE-2.0
-// 
-//    Unless required by applicable law or agreed to in writing, software
-//    distributed under the License is distributed on an "AS IS" BASIS,
-//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//    See the License for the specific language governing permissions and
-//    limitations under the License.
-
-using System.Collections;
-using System.Collections.Generic;
-
-namespace Finbuckle.MultiTenant
-{
-    /// <summary>
-    /// Holds data for configuring an InMemoryStore.
-    /// </summary>
-    public class InMemoryStoreOptions
-    {
-        public string DefaultConnectionString { get; set; }
-        public TenantConfiguration[] TenantConfigurations { get; set; }
-
-        public class TenantConfiguration
-        {
-            public string Id { get; set; }
-            public string Identifier { get; set; }
-            public string Name { get; set; }
-            public string ConnectionString { get; set; }
-            public Dictionary<string, string> Items { get; set; }
-        }
-    }
-}
