+++ src/Finbuckle.MultiTenant.AspNetCore/Internal/MultiTenantAuthenticationService.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Http;
