+++ src/Finbuckle.MultiTenant.Core/Abstractions/IMultiTenantStore.cs
@@ -12,7 +12,6 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-using System;
 using System.Threading.Tasks;
 
 namespace Finbuckle.MultiTenant
