+++ src/Finbuckle.MultiTenant.Core/Stores/InMemoryStore.cs
@@ -16,8 +16,6 @@ using System;
 using System.Collections.Concurrent;
 using System.Linq;
 using System.Threading.Tasks;
-using Finbuckle.MultiTenant.Core;
-using Microsoft.Extensions.Logging;
 
 namespace Finbuckle.MultiTenant.Stores
 {
