+++ test/Finbuckle.MultiTenant.AspNetCore.Test/Finbuckle.MultiTenant.AspNetCore.Test.csproj
@@ -5,12 +5,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Include="testsettings*.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.*" />
     <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="[2.1.0,3)" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="[2.1.0,3)" />
