+++ test/Finbuckle.MultiTenant.Core.Test/testsettings_duplicates.json
