+++ test/Finbuckle.MultiTenant.Core.Test/testsettings_duplicates_mixed_case.json
