+++ test/Finbuckle.MultiTenant.EntityFrameworkCore.Test/Extensions/MultiTenantBuilderExtensionsShould.cs
@@ -1,0 +1,60 @@
+//    Copyright 2018 Andrew White
+// 
+//    Licensed under the Apache License, Version 2.0 (the "License");
+//    you may not use this file except in compliance with the License.
+//    You may obtain a copy of the License at
+// 
+//        http://www.apache.org/licenses/LICENSE-2.0
+// 
+//    Unless required by applicable law or agreed to in writing, software
+//    distributed under the License is distributed on an "AS IS" BASIS,
+//    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//    See the License for the specific language governing permissions and
+//    limitations under the License.
+
+using Microsoft.Extensions.DependencyInjection;
+using Xunit;
+using Finbuckle.MultiTenant;
+using Finbuckle.MultiTenant.Stores;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Data.Sqlite;
+using System.Data.Common;
+
+public class MultiTenantBuilderExtensionsShould
+{
+    private DbContextOptions _options;
+    private DbConnection _connection;
+
+    public MultiTenantBuilderExtensionsShould()
+    {
+        _connection = new SqliteConnection("DataSource=:memory:");
+        _options = new DbContextOptionsBuilder()
+                .UseSqlite(_connection)
+                .Options;
+
+    }
+    [Fact]
+    public void AddEFCoreStore()
+    {
+        var services = new ServiceCollection();
+        var builder = new FinbuckleMultiTenantBuilder(services);
+        builder.WithStaticStrategy("initech").WithEFCoreStore<TestEFCoreStoreDbContext, TestTenantInfoEntity>();
+        var sp = services.BuildServiceProvider().CreateScope().ServiceProvider;
+
+        var resolver = sp.GetRequiredService<IMultiTenantStore>();
+        Assert.IsType<MultiTenantStoreWrapper<EFCoreStore<TestEFCoreStoreDbContext, TestTenantInfoEntity>>>(resolver);
+    }
+
+    [Fact]
+    public void AddEFCoreStoreWithExistingDbContext()
+    {
+        var services = new ServiceCollection();
+        var builder = new FinbuckleMultiTenantBuilder(services);
+        services.AddDbContext<TestEFCoreStoreDbContext>(o => o.UseSqlite("DataSource=:memory:"));
+        builder.WithStaticStrategy("initech").WithEFCoreStore<TestEFCoreStoreDbContext, TestTenantInfoEntity>();
+        var sp = services.BuildServiceProvider().CreateScope().ServiceProvider;
+
+        var resolver = sp.GetRequiredService<IMultiTenantStore>();
+        Assert.IsType<MultiTenantStoreWrapper<EFCoreStore<TestEFCoreStoreDbContext, TestTenantInfoEntity>>>(resolver);
+    }
+}
